Fix an exception while using the FinitePool.
