+++ src/org/bobstuff/util/Pool.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package android.util;
+package org.bobstuff.util;
 
 /**
  * @hide
