+++ src/org/bobstuff/util/Poolable.java
@@ -14,14 +14,12 @@
  * limitations under the License.
  */
 
-package android.util;
+package org.bobstuff.util;
 
 /**
  * @hide
  */
-public interface PoolableManager<T extends Poolable<T>> {
-    T newInstance();
-
-    void onAcquired(T element);
-    void onReleased(T element);
+public interface Poolable<T> {
+    void setNextPoolable(T element);
+    T getNextPoolable();
 }
