+++ src/org/bobstuff/util/SynchronizedPool.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package android.util;
+package org.bobstuff.util;
 
 /**
  *
